requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sse-starlette==1.6.5

httpx==0.25.2
PyGithub==1.59.1
requests==2.31.0

docker==6.1.3

openai==1.3.7
anthropic==0.7.7

pydantic==2.5.0
pydantic-settings==2.1.0

aiofiles==23.2.1

structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

python-dotenv==1.0.0

langgraph==0.0.40
langsmith==0.0.69
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1

jinja2==3.1.2 